Long time no sync

+++ code/simple_reader.py
@@ -4,7 +4,9 @@ import os
 import base64
 import pickle
 import tensorflow as tf
+from keras.preprocessing.sequence import pad_sequences
 from keras.preprocessing.text import Tokenizer
+from keras.utils import to_categorical
 from google.protobuf.json_format import MessageToJson
 from gensim.models import doc2vec
 
@@ -12,39 +14,40 @@ from gensim.models import doc2vec
 # https://machinelearningmastery.com/prepare-text-data-deep-learning-keras/
 
 
-def title_generator(collection, batch_size):
-    i = 0
-    titles = []
-    for record in collection.find({}):
-        title = record['metadata']['metadata']['title'].encode('utf-8')
-        titles.append(title)
-    try:
-        with open('tokenizer.pickle', 'rb') as handle:
-            t = pickle.load(handle)
-    except IOError:
-        t = Tokenizer()
-        t.fit_on_texts(titles)
-        max_len = max(map(lambda x: len(x.split(' ')), titles))
-        np.save('max_seq_length.npy', max_len)
+def get_titles(collection):
+    if (not os.path.exists('titles.npy')):
+        titles = []
+        for record in collection.find({}):
+            title = record['metadata']['metadata']['title'].encode('utf-8')
+            titles.append(title)
+        np.save('titles.npy', titles)
+    else:
+        titles = np.load('titles.npy')
+    return titles
+
+
+def pickle_tokenizer(collection):
+    titles = get_titles(collection)
+    if (not os.path.exists('tokenizer.pickle')):
+        tokenizer = Tokenizer(num_words=10000)
+        tokenizer.fit_on_texts(titles)
         with open('tokenizer.pickle', 'wb') as handle:
-            pickle.dump(t, handle, protocol=pickle.HIGHEST_PROTOCOL)
-    print batch_size
-    while (True):
-        i %= len(titles) // batch_size
-        print len(titles[i * batch_size: min(len(titles), (i + 1) * batch_size)])
-        encoded_titles = t.texts_to_matrix(titles[i * batch_size: min(len(titles), (i + 1) * batch_size)])
-        i += 1
-        yield encoded_titles
+            pickle.dump(tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
+    else:
+        with open('tokenizer.pickle', 'rb') as handle:
+            tokenizer = pickle.load(handle)
+    return tokenizer
 
 
 def extract_relevant_info(json_parsed_record, which_generator):
-    video_id = base64.b64decode(json_parsed_record['features']['feature']['video_id']['bytesList']['value'][0])
+    features = json_parsed_record['features']['feature']
+    video_id = base64.b64decode(features['video_id']['bytesList']['value'][0])
     try:
-        labels = json_parsed_record['features']['feature']['labels']['int64List']['value']
+        labels = features['labels']['int64List']['value']
     except KeyError:
         labels = []
-    mean_audio = json_parsed_record['features']['feature']['mean_audio']['floatList']['value']
-    mean_rgb = json_parsed_record['features']['feature']['mean_rgb']['floatList']['value']
+    mean_audio = features['mean_audio']['floatList']['value']
+    mean_rgb = features['mean_rgb']['floatList']['value']
     parsed_record = {
         "video_id": video_id,
         "labels": labels,
@@ -63,16 +66,23 @@ def generator(location, which_generator, batch_size):
         protobuf_records = filter(lambda x: 'test' in x, protobuf_records)
     elif (which_generator == 3):
         protobuf_records = filter(lambda x: 'validate' in x, protobuf_records)
-    protobuf_records = map(lambda x: os.path.join(location, x), protobuf_records)
+    protobuf_records = map(
+        lambda x: os.path.join(location, x),
+        protobuf_records
+    )
     i = 0
     for j in range(len(protobuf_records) // batch_size):
         print(j, "out of", len(protobuf_records) // batch_size, "done")
         i %= batch_size
         json_parsed_records = []
-        for record in protobuf_records[i * batch_size : (i + 1) * batch_size]:
+        for record in protobuf_records[i * batch_size:(i + 1) * batch_size]:
             for record_ in tf.python_io.tf_record_iterator(record):
-                json_parsed_record = json.loads(MessageToJson(tf.train.Example.FromString(record_)))
-                json_parsed_records.append(extract_relevant_info(json_parsed_record, which_generator))
+                json_parsed_record = json.loads(
+                    MessageToJson(tf.train.Example.FromString(record_))
+                )
+                json_parsed_records.append(
+                    extract_relevant_info(json_parsed_record, which_generator)
+                )
         i += 1
         if (len(json_parsed_records) == 0):
             continue
@@ -91,82 +101,104 @@ def rankAndSelectComments(comments, select):
 
 
 def filled_metadata(metadata):
-    commentCount =  metadata.get('commentCount', 0)
+    commentCount = metadata.get('commentCount', 0)
     viewCount = metadata.get('viewCount', 0)
-    favoriteCount =  metadata.get('favoriteCount', 0)
+    favoriteCount = metadata.get('favoriteCount', 0)
     dislikeCount = metadata.get('dislikeCount', 0)
     likeCount = metadata.get('likeCount', 0)
     return [commentCount, viewCount, favoriteCount, dislikeCount, likeCount]
 
 
-def mongoDBgenerator(collection, d2vmodel, numComments, which, batch_size, validation_ratio, use_audio, use_video
-        , use_desc, use_metadata, use_comments, use_channel, use_titles=False):
-    print("Found", collection.find({"which":which}).count(), "records")
-    print batch_size
+def mongoDBgenerator(
+    collection,
+    d2vmodel,
+    numComments,
+    which,
+    batch_size,
+    validation_ratio,
+    use_titles
+):
+    # print "Found {} records".format(collection.find({"which":which}).count())
     if (use_titles):
-        title_gen = title_generator(collection, batch_size)
+        tokenizer = pickle_tokenizer(collection)
+        vocab_size = tokenizer.num_words
     while True:
-        i = 0
-        X_metadata = []
-        X_audio = []
-        X_video = []
-        X_desc = []
-        X_comments = [ [] for _ in range(numComments)]
-        X_channel = []
+        X_metadata, X_audio, X_video, X_desc, X_channel =\
+            [[] for _ in xrange(5)]
+        X_comments = [[] for _ in range(numComments)]
         Y = []
         for record in collection.find({'which': which}):
-            metadata = record['metadata']['metadata']['statistics']
-            if len(record['metadata']['metadata']['comments']) < numComments:
+            metadata = record['metadata']['metadata']
+            metadata_stats = metadata['statistics']
+            if len(metadata['comments']) < numComments:
                 continue
-            metadata_records = filled_metadata(metadata) + [record['metadata']['metadata']['categoryId']]
+            metadata_records = filled_metadata(metadata_stats) +\
+                [metadata['categoryId']]
             metadata_records = np.array([int(x) for x in metadata_records])
+            if (not use_titles):
+                Y.append(d2vmodel.infer_vector(metadata['title'].split(' ')))
+            else:
+                title = metadata['title']
+                Y_ = pad_sequences(
+                    tokenizer.texts_to_sequences([title]),
+                    maxlen=15
+                )
+                Y.append(np.array(map(
+                    lambda x: map(
+                        lambda y: to_categorical(y, vocab_size),
+                        x
+                    ),
+                    Y_
+                )))
+                del Y_
             X_metadata.append(metadata_records)
             # Add metadata to training data point
-            audio_data = record['avdata']['mean_audio']
-            X_audio.append(audio_data)
+            X_audio.append(record['avdata']['mean_audio'])
             # Add audio features
-            video_data = record['avdata']['mean_rgb']
-            X_video.append(video_data)
+            X_video.append(record['avdata']['mean_rgb'])
             # Add video features
-            description = d2vmodel.infer_vector(record['metadata']['metadata']['description'].split(' '))
-            X_desc.append(description)
+            X_desc.append(
+                d2vmodel.infer_vector(metadata['description'].split(' '))
+            )
             # Add description features
-            selectedComments = rankAndSelectComments(record['metadata']['metadata']['comments'], numComments)
-            for i in range(numComments):
-                X_comments[i].append(d2vmodel.infer_vector(selectedComments[i].split(' ')))
+            selectedComments = rankAndSelectComments(
+                metadata['comments'],
+                numComments
+            )
+            for j in range(numComments):
+                X_comments[j].append(
+                    d2vmodel.infer_vector(selectedComments[j].split(' '))
+                )
             # Add comment based features
-            channelTitle = d2vmodel.infer_vector(record['metadata']['metadata']['channelTitle'].split(' '))
-            X_channel.append(channelTitle)
+            X_channel.append(
+                d2vmodel.infer_vector(metadata['channelTitle'].split(' '))
+            )
             # Add channel title features
-            if (not use_titles):
-                Y.append(d2vmodel.infer_vector(record['metadata']['metadata']['title'].split(' ')))
-            else:
-                Y = title_gen.next()
-            if len(Y) == batch_size:
-                X = []
-                if (use_metadata):
-                    X.append(np.array(X_metadata))
-                if (use_audio):
-                    X.append(np.array(X_audio))
-                if (use_video):
-                    X.append(np.array(X_video))
-                if (use_desc):
-                    X.append(np.array(X_desc))
-                if (use_comments):
-                    X += [ np.array(commentVec) for commentVec in X_comments ]
-                if (use_channel):
-                    X.append(np.array(X_channel))
+            if len(X_channel) == batch_size:
+                X = [
+                    np.array(X_metadata),
+                    np.array(X_audio),
+                    np.array(X_video),
+                    np.array(X_desc)
+                ]
+                # Return all data from generator, handle in model
+                X += [np.array(commentVec) for commentVec in X_comments]
+                # if (use_channel):
+                X.append(np.array(X_channel))
                 Y = np.array(Y)
+                if use_titles:
+                    Y = np.squeeze(Y, axis=1)
                 # Split into train and validation if training
                 if which == 1:
                     splitPoint = int(validation_ratio * len(Y))
-                    X_train = [ x[splitPoint:] for x in X]
-                    X_val = [ x[:splitPoint] for x in X]
+                    X_train = [x[splitPoint:] for x in X]
+                    X_val = [x[:splitPoint] for x in X]
                     yield (X_train, Y[splitPoint:]), (X_val, Y[:splitPoint])
                 else:
                     yield X, Y
-                X_metadata, X_audio, X_video, X_desc, X_channel = [], [], [], [], []
-                Y, X_comments = [], [ [] for _ in range(numComments) ]
+                X_metadata, X_audio, X_video, X_desc, X_channel =\
+                    [[] for _ in xrange(5)]
+                Y, X_comments = [], [[] for _ in range(numComments)]
         # Don't loop indefinitely if test data
         if which != 1:
             break
